RF: Minimalist datalad.dataset.gitrepo interface
Basically just the pieces needed to point to a local repository
and execute call_git*() methods.
